inicio alterações estrutura
