+++ backend/controller/log_controller.py
@@ -1,7 +1,7 @@
 import sys
 sys.path.append('.')
 #from backend.dao_txt.log import listar_log
-from backend.dao_bd.log import *
+from backend.dao_bd.log_dao import *
 
 
 def listar_logs() -> list:
