+++ backend/controller/marketplace_controller.py
@@ -1,8 +1,8 @@
 import sys
 sys.path.append('.')
-from backend.dao_txt.marketplace import criar_marketplace, listar_marketplace
-from backend.dao_bd.log import *
-from backend.dao_bd.marketplace import *
+from backend.dao_txt.marketplace_dao_txt import criar_marketplace, listar_marketplace
+from backend.dao_bd.log_dao import *
+from backend.dao_bd.marketplace_dao import *
 
 
 def criar_marketplaces(nome:str, descricao:str) -> None:
